hipd: change a variable type from size_t to int. The size_t is an unsigned one, and using it for a variable that * takes the (possibly negative) retval of a call AND * is compared to zero makes no sense.

+++ hipd/dh.c
@@ -24,7 +24,7 @@ static void hip_regen_dh_keys(uint32_t bitmask);
  */
 int hip_insert_dh(uint8_t *buffer, int bufsize, int group_id)
 {
-    size_t res;
+    int res;
     DH *tmp;
 
     /*
